% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{BestWeightsByLambda}
\alias{BestWeightsByLambda}
\title{Find best weights}
\usage{
BestWeightsByLambda(performance, lambdas, alpha = 0, glmData, nBack)
}
\arguments{
\item{performance}{all results from regularized regression after running RegularizedRegression}

\item{lambdas}{list of lambda parameters that was used to run RegularizedRegression}

\item{glmData}{trial-by-trial subject responses}

\item{nBack}{number of trials to go back in history. Default is 1.}
}
\description{
Chooses best weights out of all weights generated for each lambda
of regularized regression.
Then, compute glm coefficients on full dataset using the best lambda
}

