% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{BiasAfterRandomization}
\alias{BiasAfterRandomization}
\title{Compute history weights after either shuffling trial order or shuffling}
\usage{
BiasAfterRandomization(rawData, randomizationType = 1, alpha = 0,
  nBack = 1, B = 10)
}
\description{
responses to each trial. to computing by either scrambling the order of
trials, or by generating responses randomly. This procedure should
decrease the history weights to 0. If we run this function many times we
will get confidence intervals for history biases
}

