% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{BuilDataForGLM}
\alias{BuilDataForGLM}
\title{BuldDataForGLM
*** FUNCTION to prepare all necessary columns and fill in data for the GLM analysis
*** Previous success and previous failure (history terms) columns are also constructed here}
\usage{
BuilDataForGLM(rawData, nHistoryBack, nDummyParams, successColName, failColName)
}
\description{
BuldDataForGLM
*** FUNCTION to prepare all necessary columns and fill in data for the GLM analysis
*** Previous success and previous failure (history terms) columns are also constructed here
}

