% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_PreprocessData.R
\name{ClassifyRawData}
\alias{ClassifyRawData}
\title{Assign "Conditions"
Condition 1: No bias condition. fail/switch and succeed/stay probabilities are set to 50%
Condition 2: fail/switch probability set at 80%, succeed/stay at 50%. !!!IMPORTANT. This is to induce fail/stay bias, because when we switch after fail, we end up on same side where fail was.
Condition 3: fail/switch probability set at 50%, succeed/stay at 80%. This is to induce success/stay bias.
Condition 7: succeed/stay probability set to 50%, fail/switch is set to 20%. This is to induce fail/switch bias, because after fail stimulus will likely to remain on the opposite side of fail, inviting the subbject to switch side.
Condition 8: succeed/stay probability=20%, fail/switch=50%. Lots of switching after success to induce success/switch bias.
Condition 9: This is same as Condition 2, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=50%, fail/switch=80%.
Condition 10: This is same as Condition 3, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=80%, fail/switch=50%.
Condition 11: This is same as Condition 7, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=50%, fail/switch=20%.
Condition 12: This is same as Condition 8, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=20%, fail/switch=50%.
Condition 4: Stimulus diameter is 6??, eccentricity is 8??
Condition 5: Stimulus diameter is 12??, eccentricity is 12??
Condition 6: Stimulus diameter is 6??, eccentricity is 10??}
\usage{
ClassifyRawData(rawData)
}
\description{
Assign "Conditions"
Condition 1: No bias condition. fail/switch and succeed/stay probabilities are set to 50%
Condition 2: fail/switch probability set at 80%, succeed/stay at 50%. !!!IMPORTANT. This is to induce fail/stay bias, because when we switch after fail, we end up on same side where fail was.
Condition 3: fail/switch probability set at 50%, succeed/stay at 80%. This is to induce success/stay bias.
Condition 7: succeed/stay probability set to 50%, fail/switch is set to 20%. This is to induce fail/switch bias, because after fail stimulus will likely to remain on the opposite side of fail, inviting the subbject to switch side.
Condition 8: succeed/stay probability=20%, fail/switch=50%. Lots of switching after success to induce success/switch bias.
Condition 9: This is same as Condition 2, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=50%, fail/switch=80%.
Condition 10: This is same as Condition 3, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=80%, fail/switch=50%.
Condition 11: This is same as Condition 7, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=50%, fail/switch=20%.
Condition 12: This is same as Condition 8, but only low contrasts used to have similar number of failure and success. Succeed/stay probability=20%, fail/switch=50%.
Condition 4: Stimulus diameter is 6??, eccentricity is 8??
Condition 5: Stimulus diameter is 12??, eccentricity is 12??
Condition 6: Stimulus diameter is 6??, eccentricity is 10??
}

