% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{ComputeWeightsWithBestLambda}
\alias{ComputeWeightsWithBestLambda}
\title{Find the best lambda}
\usage{
ComputeWeightsWithBestLambda(performance, lambdas, glmData, nBack)
}
\description{
One which produces smallest log likelihood.
Then, compute glm coefficients on full dataset using the best lambda
}

