% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{ConditionLabels}
\alias{ConditionLabels}
\title{Labels assigned to each experimental condition}
\usage{
ConditionLabels(variable, value)
}
\description{
On 09 Jan 2014, swapped labels for fail/stay and fail/success conditions.
Initially, Condition 2 was fail/switch, but then realized that switching trial
position after a failure actually encourages to stay on the same side where failure
happened. Conversly, fail/stay (stim presented on same side) actuallly encourages
to switch choice on next trial cause failure happened on the opposite side.
}

