% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{CorrelateAgeAndBias}
\alias{CorrelateAgeAndBias}
\title{Compute correlation between age and bias}
\usage{
CorrelateAgeAndBias(sbjInfo, biases)
}
\description{
Compute correlation between age and bias
}

