% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{CV.glmnet}
\alias{CV.glmnet}
\title{**************************
*** stub for glmnet cross-validation that accepts a formula}
\usage{
CV.glmnet(formula, data, subset, na.action, ...)
}
\description{
**************************
*** stub for glmnet cross-validation that accepts a formula
}

