% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{FitProbit}
\alias{FitProbit}
\title{Fit probit psychometric function}
\usage{
FitProbit(dat, fitWithLapseRate = FALSE)
}
\arguments{
\item{dat}{prorotion of rightward responses computed using the function ProportionRightwardResponses}

\item{fitWithLapseRate}{if TRUE, fits probit that takes into account the lapse rate,
which is computed as a proportion of incorrect responses to the highest stimulus intensity (this is safe to do for
our experiment, cause we deliberately selected highest stimulus intensity to be 100% of time detectable). Lapse rate is
part of `dat`.}
}
\description{
Fits probit (cummmulative Gaussian) to proportion of rightward responses. The best fit after Weibull (with
log transformed contrast values). Can fit with or without lapse rate.
}
\examples{
oneRun <- droplevels(subset(glmData, SubjectID=='s025' & SessionID==3))
pRight <- ProportionRightwardResponses(oneRun)
probitModel <- FitProbit(pRight)
range <- c(-0.08,0.08)
predicted <- PredVals(probitModel, xvar="Contrast", yvar="pRightCorrect", type="response", xrange=range)
ggplot(pRight, aes(x=Contrast, y=pRightCorrect))+
 ylim(0,1) +
 scale_x_continuous(limits=range, breaks=seq(range[1],range[2], 0.02)) +
 geom_point() +
 geom_line(data=predicted, aes(x=Contrast, y=pRightCorrect))
}

