% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{FitRegulirizedLogisticRegression}
\alias{FitRegulirizedLogisticRegression}
\title{Fit model parameters using regulirized logistic regression (glmnet)}
\usage{
FitRegulirizedLogisticRegression(glmData, lambdas, alpha = 0, B,
  fitHistoryBias = TRUE)
}
\description{
Fit model parameters using regulirized logistic regression (glmnet)
}

