% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{FitWeibull}
\alias{FitWeibull}
\title{Fit Weibull function}
\usage{
FitWeibull(dat)
}
\description{
Fit Weibull without log transforming contrast values
AVOID using this because the fit is poor
}

