% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{FitWeibullLogContrast}
\alias{FitWeibullLogContrast}
\title{Fit Weibull after log transforming contrast values}
\usage{
FitWeibullLogContrast(data)
}
\description{
This produces the best fit out of all psychometric curves
NOTE: when contrast is negative, an easier approach is to fit probit
Tried Gumbel, but it didn't provide a good fit.
}

