% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{ForestPlot}
\alias{ForestPlot}
\title{Plot model parameters as a distance from vertical}
\usage{
ForestPlot(plotData, xlims = c(-3, 3), xbreaks = c(-2, -1, 0, 1, 2),
  figureWidth = 4.820225, figureHeight = 7.896552, plotMeans = TRUE,
  plotMeanPointsHollow = TRUE, plotEachWeight = TRUE,
  plot95ConfInt = FALSE, plotBackground = FALSE, plotByCondition = TRUE,
  plotByParameter = FALSE)
}
\description{
Forest plot provides convenient way of plotting model parameters to see how far
they deviate from 0
}

