% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{Glmnet}
\alias{Glmnet}
\title{**************************
*** stub for glmnet that accepts a formula
Adapted from a book (NEED TO CHECK THE REF)}
\usage{
Glmnet(formula, data, subset, na.action, ...)
}
\description{
**************************
*** stub for glmnet that accepts a formula
Adapted from a book (NEED TO CHECK THE REF)
}

