% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_PlottingFunctions.R
\name{HistoryAdaptationScatterplot}
\alias{HistoryAdaptationScatterplot}
\title{History scatterplot that shows how history biases adapt}
\usage{
HistoryAdaptationScatterplot(weights, figureWidth = 15.516854,
  figureHeight = 5.689655, conditionsToPlot, sourceCondition = NA,
  plotErrorBars = FALSE, plotByCondition = FALSE,
  plotSubjectMeans = FALSE, plotSubjectNames = FALSE, plotLegend = TRUE)
}
\description{
Formerly "HistoryMigrationScatterplot"
Scatterplot of choice history weights grouped by Education
This is a plot for the manuscript
}

