% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{LapseRateFromHighestContrast}
\alias{LapseRateFromHighestContrast}
\title{Compute lapse rate from highest contrast intensity}
\usage{
LapseRateFromHighestContrast(rightwardResponses)
}
\arguments{
\item{rightwardResponses}{data in the format of proportion of rightward responses (also called pcRightSummary) for one subject. It can either be average across many runs or one run. The function doesn't care about it.}
}
\value{
lapse rate as proportion of errors at the highest contrast intensity
}
\description{
Lapses occur due to inattention and are not related to decision making processes
in the brain. Here, the lapse rate is estimated by finding the strongest contrast intensity
 - when missing stimulus is likely to be due to inattention - and computing the error rate
 at that stimulus intensity
}
\details{
This lapse rate estimation approach is based on common sense, experience and supported by:
Prins, N. (2012). The psychometric function: the lapse rate revisited. Journal of Vision, 12(6). http://doi.org/10.1167/12.6.25

NOTE: If your data does not strong contrast intensities, using this method of lapse rate estimation is not recommended.
In such a case, it is better to use
}

