% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_PlottingFunctions.R
\name{PlotContrastAndHistoryWeights}
\alias{PlotContrastAndHistoryWeights}
\title{Plot both contrast and choice history weights}
\usage{
PlotContrastAndHistoryWeights(regularizedWeights, plotByCondition = T,
  legendPos = c(0.15, 0.75), plotInColor = FALSE, showAxisLabels = TRUE,
  plotFigure = TRUE, figureWidth = 6.382023, figureHeight = 8.678161)
}
\description{
A beautiful plot of all weights (contrast and history). Contrast weights are
connected together, but history weights are not.
}

