% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{PlotContrastAndHistoryWeights_OLD}
\alias{PlotContrastAndHistoryWeights_OLD}
\title{Plot both contrast and choice history weights}
\usage{
PlotContrastAndHistoryWeights_OLD(regularizedWeights, plotByCondition = T,
  legendPos = c(0.15, 0.75), plotInColor = F, showAxisLabels = T,
  figureWidth = 6.382023, figureHeight = 8.678161)
}
\description{
A beautiful plot of all weights (contrast and history). Contrast weights are
connected together, but history weights are not.
}

