% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{PlotContrastPairs}
\alias{PlotContrastPairs}
\title{Function to plot stimulus intensities during switching or staying behaviours}
\usage{
PlotContrastPairs(contrastPairs, plotTitle)
}
\description{
Function to plot stimulus intensities during switching or staying behaviours
}

