% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_PlottingFunctions.R
\name{PlotPsychometricCurves}
\alias{PlotPsychometricCurves}
\title{Plot psychometric curves}
\usage{
PlotPsychometricCurves(inputData, figureWidth = 21.37078,
  figureHeight = 9.034483, plotByCondition = T, plotErrorBars = T,
  findAsymptotes = F, collapseByRun = T, lineColor = "#c6dbef",
  dotColor = "#09519c", nrow = 3)
}
\description{
Plot the proportion of right responses.
Also fit Weibull and show the fit
}

