% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_PlottingFunctions.R
\name{PlotSensitivityDecline}
\alias{PlotSensitivityDecline}
\title{Show decline in visual sensitivity from choice history biases}
\usage{
PlotSensitivityDecline(simThAndSlope, whatToPlot = "slope")
}
\arguments{
\item{simThAndSlope}{simulated threshold and slope computed using function NoBiasVsSubjectBias (see example below).}

\item{whatToPlot}{either decline of slope or threshold can be plotted}
}
\description{
Displays decline in visual sensitivity as a result of choice history biases.
It can show reduction in contrast sensitivity (75% contrast threshold) or
change in slope of psychometric function (in case of probit, smaller slope
means less sensitivity).
}
\examples{
load('20150919_allWeights_RIKEN_UCL_Stanford_cond1n13.RData',verbose=TRUE)
regWeights <- droplevels(subset(allWeights, Regularized=='yes'))
oneSbjWeights <- droplevels(subset(regWeights, SubjectID \%in\% c('s001', 's002', 's003', 's009', 's008', 's007')))
thAndSlope <- NoBiasVsSubjectBias(regWeights, nTrialsPerContrast=10, B=5) # B=5 is for fast simulation, for meaningful results use B>500.
PlotSensitivityDecline(thAndSlope) #
}

