% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_PlottingFunctions.R
\name{PlotSparklineOfRun}
\alias{PlotSparklineOfRun}
\title{Plot whole run as Tuft's sparklines of left/right choice and responses}
\usage{
PlotSparklineOfRun(runData, plotResults = F, plotInColor = T)
}
\description{
Plot whole run as Tuft's sparklines of left/right choice and responses
}

