% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{PlotThVsBias}
\alias{PlotThVsBias}
\title{Plot contrast threshold vs subject bias}
\usage{
PlotThVsBias(inputData, modelWeights, fitWithLapseRate = TRUE,
  conditionsToPlot = 1)
}
\arguments{
\item{inputData}{trial-by-trial data of glmData type}

\item{modelWeights}{regularized model weights of subjects that were computed on `inputData`}

\item{fitWithLapseRate}{if TRUE, fits psychometric curves by considering the lapse rate}

\item{conditionsToPlot}{conditions to plot. Default is `1`}
}
\description{
Shows if contrast threshold correlates with with history weights.
For example, subjects with higher visual sensitivity might
have smaller biases because they have to rely less on priors and more on sensory evidence
}
\details{
TODO: Function also plots relationship between biases and RT. Need to incorporate this as a switch in parameters (e.g., whatToPlot=c('th', 'rt'))
}
\examples{
PlotThVsBias(glmData, regWeights, conditionsToPlot=c(1,13))

}

