% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{PlotVIFs}
\alias{PlotVIFs}
\title{Function to plot Variance Inflation Factor (VIF)}
\usage{
PlotVIFs(allWeights)
}
\description{
Function to plot Variance Inflation Factor (VIF)
}

