% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{PrepareRawData}
\alias{PrepareRawData}
\title{Tidy up rawData
Previous success and previous failure (history terms) columns are also constructed here}
\usage{
PrepareRawData(rawData)
}
\description{
Tidy up rawData
Previous success and previous failure (history terms) columns are also constructed here
}

