% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chb_Simulations.R
\name{PrepareTrials}
\alias{PrepareTrials}
\title{Prepare a dataframe with a given contrast intensity}
\usage{
PrepareTrials(contrast, subjectID = "ERR", Condition = -1)
}
\description{
Makes all necessary columns that typically correspond to a trial.
However, this function returns four trials, by splitting
it into 2 VisualFields (left/right) and two drifts (left/right)
It is then possible to replicate rows of this dataframe
typically used to simulations
}

