% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{ProbitThreshold}
\alias{ProbitThreshold}
\title{Compute threshold of psychometric function which was fitted with probit}
\usage{
ProbitThreshold(p, model)
}
\description{
Takes in a glm object and probability at which to estimate the threshold
}

