% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{RegularizedRegression}
\alias{RegularizedRegression}
\title{Fit model parameters using regularized regression}
\usage{
RegularizedRegression(glmData, fitWithLapseRate = FALSE, lambdas, alpha = 0,
  B, fitHistoryBias = TRUE)
}
\description{
Uses regulirized logistic regression (glmnet)
TODO: describe fully how training and validation done to find the best lambda
}

