% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_SimulationFunctions.R
\name{SensitivtyDeclineMatrix}
\alias{SensitivtyDeclineMatrix}
\title{Compute decline in sensitivity as a function of choice history biases}
\usage{
SensitivtyDeclineMatrix(regWeights, failWeightsToSim, successWeightsToSim,
  nTrialPerContrast, B = 30, setLRBiasToZero = TRUE)
}
\arguments{
\item{regWeights}{subject weights}

\item{failWeightsToSim}{a matrix of failure weights}

\item{successWeightsToSim}{and matrix of success weights}

\item{nTrialPerContrast}{number of trials per contrast intensity}

\item{B}{number of simulations. A good number to start with is 30. Larger is better. T}

\item{setLRBiasToZero}{either L/R bias of the model as it is or set subjects' L/R bias to zero}
}
\value{
Returns a data frame containing threshold and slope values of each simulation done with all combinations of success and fail weights
}
\description{
Takes subject model weights, computes mean subject weights and then take the average
across all those subject means which results in one set of weights, which is average across
all subjects. Basically, it is like running a simulation with one "average" subject, who is the
average of all subjects. This approach saves time because this simulation take a while to run.
}
\details{
This function stores thresholds and slopes of all simulations. To plot the results use
PlotDeclineMatrix function.
}
\examples{
# This will run a massive simulation of subject responses
by using 13x13=169 matrix of fail and success weights and each
matrix (i.e., 169 simulations) will be run 50 times.
SensitivtyDeclineMatrix(regWeights, failWeightsToSim = seq(-2,2, length.out = 13),
                        successWeightsToSim = seq(-2,2, length.out = 13), nTrialsPerContrast=50, B=50)
# A quicker simulation
SensitivtyDeclineMatrix(regWeights, failWeightsToSim = c(-2, -1,0,1, 2),
                       successWeightsToSim = c(-2, -1,0,1, 2), nTrialsPerContrast=30, B=30)

}

