% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_SimulationFunctions.R
\name{SimulateNoBiasVsSubjectBias}
\alias{SimulateNoBiasVsSubjectBias}
\title{Estimate slope and threshold with and without history biases}
\usage{
SimulateNoBiasVsSubjectBias(sbjWeights, nTrialsPerContrast = 100, B = 30,
  setLRBiasToZero = TRUE)
}
\arguments{
\item{sbjWeights}{regularized subject weights}

\item{nTrialsPerContrast}{number of trials for each contrast}

\item{B}{number of simulations}

\item{setLRBiasToZero}{removes L/R bias (intercept) from all simulations, including from subject's weights}
}
\description{
Run simulation using either subjects' natural history biases or without
history biases (success, failure and L/R biases are set to 0).
It sets success, failure and L/R biases of each subject
to 0 when simulating no-bias responses and leaves them intact when simulating
using subject's natural biases. The "IsSubject" field indicates if simulates
has subject's natural biases (TRUE) or has biases removed (FALSE).
}
\details{
Note: Simulations run using subjects' mean weights
}
\examples{
SimulateNoBiasVsSubjectBias(sbjWeights, nTrialsPerContrast = 100, B = 30)
}

