% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chb_Simulations.R
\name{SimulateOneSubject}
\alias{SimulateOneSubject}
\title{Simulates responses of one subject}
\usage{
SimulateOneSubject(subjectWeights, SubjectID = "temp", Condition = -777,
  trialsPerContrast = 100, nBack = nBack, lapseRate = 0)
}
\arguments{
\item{subjectWeights}{contrast and history bias weights including L/R bias}

\item{SubjectID}{subject code, such as 's007'}

\item{Condition}{a number that specifiec condition of simulation}

\item{trialsPerContrast}{number of trials that each contrast intensity will be simulated}

\item{nBack}{history depth. At the moment this is not functional???}
}
\description{
Simulate one subject using contrast weights, L/R bias and history weights
The simulation is computed ONLY ONCE using certain number of trials specified by
trialsPerContrast. You can call this function multiple times to estimates
confidence intervals of slope or threshold, for example
}
\examples{
SimulateOneSubject(oneSbjWeights, nTrialsPerContrast=50)
}

