% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_SimulationFunctions.R
\name{SimulateSubjectResponses}
\alias{SimulateSubjectResponses}
\title{Simulate subject responses for different set of failure weights}
\usage{
SimulateSubjectResponses(subjectWeights, failWeightsToSim = NA,
  successWeightsToSim = NA, nTrialsPerContrast = 1000,
  setLRBiasToZero = TRUE, conditionsToSimulate, lapseRate = 0)
}
\description{
Also controls setting L/R bias and success bias to 0
This function can be generalized because in current form
it only assumes that only failure biases change but success
biases remain either constant or absent (same for L/R bias)
}

