% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{SortParams}
\alias{SortParams}
\title{Sort parameters of the model in intuitive way}
\usage{
SortParams(paramNames)
}
\arguments{
\item{paramNames}{model parameter names as factor (see usage)}
}
\description{
Model parameter names can get scrambled and when plotting them
it can be difficult to understand the graph. This function
convniently groups model parameters by name such that contrast
intensities are followed by bias weights.
}
\examples{
modelWeights$Parameters <- sortParams(modelWeights$Parameters)
}

