% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{StimulusSideSequences}
\alias{StimulusSideSequences}
\title{Function to compute how long were sequences of stimuli either on L or R sides.
This is to figure out if there are long sequences, particularly when introducing biases
such as succeed/stay bias which can generate long sequences on one side}
\usage{
StimulusSideSequences(glmData)
}
\description{
Function to compute how long were sequences of stimuli either on L or R sides.
This is to figure out if there are long sequences, particularly when introducing biases
such as succeed/stay bias which can generate long sequences on one side
}

