% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{SubjectDemographics}
\alias{SubjectDemographics}
\title{Subject info}
\usage{
SubjectDemographics(getAge = TRUE, getGender = TRUE, getEducation = TRUE,
  getLab = TRUE)
}
\arguments{
\item{getAge}{if TRUE, returns subjects' age}

\item{getGender}{if TRUE, returns subjects' gender}

\item{getEducation}{if TRUE, returns subjects' education (PhD vs No-Phd)}

\item{getLab}{if TRUE, returns lab location of subjects (Stanford, UCL or RIKEN)}
}
\description{
Returns age, gender, education and lab location of subjects
}

