% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{ThAndSlope}
\alias{ThAndSlope}
\title{Get threshold and slope from glmData}
\usage{
ThAndSlope(inputData, returnMeans = TRUE, whichConditions = 1)
}
\description{
Works better with subjects' real data rather than simulated trials
For simulated trials, better to use function called ThAndSlopeForSimData
}

