% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_Simulations.R
\name{ThAndSlopeForSimData}
\alias{ThAndSlopeForSimData}
\title{Compute threshold and slope from simulated trials}
\usage{
ThAndSlopeForSimData(simTrials, fitWithLapseRate = TRUE)
}
\arguments{
\item{simTrials}{simulated trials, which can be done using the function SimulateSubjectResponses}

\item{fitWithLapseRate}{if TRUE, the probit psychometric function will be fitted using the lapse rate}
}
\description{
There is another function called ThAndSlope which works
with subjects' data rather than simulated data, because simulated
data of glmData type includes additional column called IsSubject
}

