% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{ThSlopeAndLapse}
\alias{ThSlopeAndLapse}
\title{Get threshold, slope and lapse rate from glmData}
\usage{
ThSlopeAndLapse(inputData, fitWithLapseRate = TRUE, returnMeans = TRUE,
  whichConditions)
}
\arguments{
\item{inputData}{data structured as glmData}

\item{fitWithLapseRate}{if TRUE, the Probit psychometric function is computed as $\lambda+(1-2*\lambda)Probit$, where lambda is the lapse rate. The lapse rate is computed using the function LapseRateFromHighestContrast}

\item{returnMeans}{if TRUE, computes average for each subject. Otherwise, returns run-by-run results}

\item{whichConditions}{conditions to analyse. For natural bias, include conditions 1 and 13}
}
\value{
data frame containing threshold, slope and lapse rate of each subject. If requested, can return by-run results, otherwise means of each subject
}
\description{
Works better with subjects' real data rather than simulated trials
For simulated trials,use ThAndSlopeForSimData
}
\examples{
ThSlopeAndLapse(glmData, returnMeans=TRUE, whichCondition=c(1,13))

}

