% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{WeibullThAndSlope}
\alias{WeibullThAndSlope}
\title{Estimate threshold and slope using Weibull}
\usage{
WeibullThAndSlope(p, model)
}
\description{
DON"T USE THIS. The most accurate threshold estimation is probit.
Compute threshold and slope of psychometric function which was fitted with a
Weilbull. takes in a glm object and probability at which to estimate the threshold
Adapted from "Modelling Psychophysical Data in R", p. 155
}

