% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_PreprocessData.R
\name{AssignCondition}
\alias{AssignCondition}
\title{Assign condition to sessions}
\usage{
AssignCondition(dat, SubjectID, SessionID, Condition)
}
\description{
Important: SessionID should be provided based on OriginalSessionID
}

