% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{BiasAfterRandomization}
\alias{BiasAfterRandomization}
\title{Compute weights after randomizing trials or responses}
\usage{
BiasAfterRandomization(rawData, randomizationType = 1, alpha = 0,
  nBack = 1, B = 10)
}
\description{
This is checker function to ensure that history weights do indeed
disapper after trial order is made random or subjects responses are made
random. This randomization effectively destroys history effects effectively
setting history weights to 0.
!!! CHECK IF YOU HAVE ANOTHER FUNCTION THAT DOES THIS!!!! It is a suspect, because
uses rawData as input and not glmData, which is much more clean data.
Again, compute history weights after either shuffling trial order or shuffling
responses to each trial, by either scrambling the order of
trials, or by generating responses randomly. This procedure should
decrease the history weights to 0. If we run this function many times we
will get confidence intervals for history biases
}

