% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{BuilDataForGLM}
\alias{BuilDataForGLM}
\title{Prepare data for GLM analysis}
\usage{
BuilDataForGLM(rawData, nHistoryBack, nDummyParams, successColName, failColName)
}
\description{
Build all necessary columns and fill in data for the GLM analysis
Previous success and previous failure (history terms) columns are also constructed here
}

