% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{ComputeWeightsWithBestLambda}
\alias{ComputeWeightsWithBestLambda}
\title{Find the best lambda}
\usage{
ComputeWeightsWithBestLambda(performance, lambdas, glmData, lapseRate, nBack)
}
\description{
One which produces smallest log likelihood.
Then, compute glm coefficients on full dataset using the best lambda
}

