% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{CorrelateAgeAndBias}
\alias{CorrelateAgeAndBias}
\title{Compute correlation between age and bias}
\usage{
CorrelateAgeAndBias(sbjInfo, biases)
}
\description{
Compute correlation between age and bias
}

