% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{CorrelationMatrixOfParameters}
\alias{CorrelationMatrixOfParameters}
\title{Compute correlation matrices for each parameter of the GLM model
This helps to check if there are any strong effects of multicolinearity}
\usage{
CorrelationMatrixOfParameters(glmData)
}
\description{
Compute correlation matrices for each parameter of the GLM model
This helps to check if there are any strong effects of multicolinearity
}

