% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{CV.glmnet}
\alias{CV.glmnet}
\title{Stub for glmnet cross-validation that accepts a formula}
\usage{
CV.glmnet(formula, data, subset, na.action, ...)
}
\description{
Stub for glmnet cross-validation that accepts a formula
}

