% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{FitProbit}
\alias{FitProbit}
\title{Fit probit function}
\usage{
FitProbit(dat)
}
\description{
Fit probit (cummmulative Gaussian). The best fit after Weibull (with
log transformed contrast values)
}

