% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{FitRegulirizedLogisticRegression}
\alias{FitRegulirizedLogisticRegression}
\title{Fit model parameters}
\usage{
FitRegulirizedLogisticRegression(glmData, lambdas, alpha = 0, B,
  fitHistoryBias = TRUE)
}
\description{
Uses regulirized logistic regression (glmnet)
}

