% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{FitWeibull}
\alias{FitWeibull}
\title{Fit Weibull function}
\usage{
FitWeibull(dat)
}
\description{
Fit Weibull without log transforming contrast values
AVOID using this because the fit is poor
}

