% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{Glmnet}
\alias{Glmnet}
\title{Stub for glmnet that accepts a formula}
\usage{
Glmnet(formula, data, subset, na.action, ...)
}
\description{
Adapted from a book (NEED TO CHECK THE REF)
}

