% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chb_Plotting.R
\name{HistoryScatterplotByEducation}
\alias{HistoryScatterplotByEducation}
\title{Scatterplot of choice history weights grouped by Education}
\usage{
HistoryScatterplotByEducation(weights, figureWidth = 15.516854,
  figureHeight = 5.689655, conditionsToPlot, legendPos = c(0.15, 0.75),
  plotMeansByEducation = F, plotByCondition = TRUE,
  plotSubjectMeans = FALSE, plotSubjectNames = FALSE)
}
\description{
This is a plot for the manuscript
}

