% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_PlottingFunctions.R
\name{HistoryWeightsScatterplot}
\alias{HistoryWeightsScatterplot}
\title{A scatterplot of choice history biases}
\usage{
HistoryWeightsScatterplot(weights, figureWidth = 15.516854,
  figureHeight = 5.689655, xlims = c(-3, 3), ylims = c(-3, 3),
  conditionsToPlot, plotByCondition = TRUE, plotSubjectMeans = FALSE,
  plotSubjectNames = FALSE, plotMeans = FALSE)
}
\description{
Plot 'prevFail' vs 'prevSuccess' weights in forms of scatterplot.
}

