% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{PrepareRawData}
\alias{PrepareRawData}
\title{Enrich rawData with helper columns}
\usage{
PrepareRawData(rawData)
}
\description{
Previous success and previous failure (history terms) columns are also constructed here
}

