% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{ProbitThreshold}
\alias{ProbitThreshold}
\title{Compute contrast threshold using fitted probit model}
\usage{
ProbitThreshold(p, model)
}
\arguments{
\item{p}{either a single value or an array of probabilities for which contrast thresholds will be computed}

\item{model}{a glm object of fitted probit psychometric curve}
}
\description{
Estimates contrast intensities which generate response
at probablity p (can be an array). The fitted psychometric
parameters are in model glm.
}

