% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chb_Simulations.R
\name{SimulateResponses}
\alias{SimulateResponses}
\title{Simulate responses using given model weights}
\usage{
SimulateResponses(subjectWeights, nSimulations = 10, trialsPerContrast = 50,
  Conditions)
}
\arguments{
\item{subjectWeights}{}

\item{nSimulations}{}

\item{trialsPerContrast}{}

\item{Conditions}{}
}
\description{
BETTER NOT TO USE THIS. Instead, use SimulateSubjectResponses
Compared to SimulateSubjectResponses, this function is simpler
because it only deals with given set of weights and does not
allow running different model parameters. This function
can actually be called from within SimulateSubjectResponses
but for some reason it is not done. A TODO for future.
}

