% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{StimulusSideSequences}
\alias{StimulusSideSequences}
\title{Compute length of sequences on left or right sides}
\usage{
StimulusSideSequences(glmData)
}
\description{
This is to figure out if there are long sequences, particularly when introducing biases
such as succeed/stay bias which can generate long sequences on one side
}

