% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_DataAnalysisFunctions.R
\name{SubjectDemographics}
\alias{SubjectDemographics}
\title{Subject demographics such as age and education}
\usage{
SubjectDemographics(getAge = TRUE, getGender = TRUE, getEducation = TRUE)
}
\description{
Subject demographics such as age and education
}

