% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Venus_SimulationFunctions.R
\name{ThAndSlopeForSimData}
\alias{ThAndSlopeForSimData}
\title{Compute threshold and slope from simulated trials}
\usage{
ThAndSlopeForSimData(simTrials)
}
\description{
There is another function called ThAndSlope which works
with subjects' data rather than simulated data, because simulated
data of glmData type includes additional column called IsSubject
}

