% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chb_DataAnalysis.R
\name{ThSlopeAndLapse}
\alias{ThSlopeAndLapse}
\title{Get threshold, slope and lapse rate from glmData}
\usage{
ThSlopeAndLapse(inputData, returnMeans = TRUE, whichConditions)
}
\arguments{
\item{inputData}{data structured as glmData}

\item{returnMeans}{if TRUE, computes average for each subject. Otherwise, returns by-run results}

\item{whichConditions}{conditions to analyse. For natural bias, include conditions 1 and 13}
}
\value{
data frame containing threshold, slope and lapse rate of each subject. If requested, can return by-run results, otherwise means of each subject
}
\description{
Works better with subjects' real data rather than simulated trials
For simulated trials,use ThAndSlopeForSimData
}
\examples{
ThSlopeAndLapse(glmData, returnMeans=TRUE, whichCondition=c(1,13))
}

